pyproject.toml
[tool.poetry]
name = "celery-fiber"
version = "0.1.0"
description = ""
authors = ["Patrick Allen <prallen90@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
celery = "^4.3"
flask = "^1.1"
requests = "^2.22"
eventlet = "^0.25.1"
aiohttp = "^3.6"
python-socketio = "^4.3"
sanic = "^19.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
